# apps/api/openapi.snapshot.json
{
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "description": "Machine-readable error code (e.g. 'invalid_body', 'unauthenticated').",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "JWT (dashboard sessions) or `furan_pat_*` (SDK/CI). Both arrive on `Authorization: Bearer ...`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Visual regression testing API. See https://github.com/qasecret/furan-monorepo for self-host docs.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Furan API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs/{id}/events": {
      "get": {
        "description": "Long-lived response (`text/event-stream`). Connection closes when the client disconnects.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "description": "SSE frames in the format `event: progress\\ndata: <JSON>\\n\\n`. Keep-alive comments `:\\n\\n` sent every 15s.",
                  "type": "string"
                }
              }
            },
            "description": "SSE stream of run events"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not a project member"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Run not found"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "Server-Sent Events stream for a run. Emits `event: progress` frames with JSON payloads carrying `{ event, runId, ...details }`. Sends a `:` keep-alive every 15s.",
        "tags": [
          "sse"
        ]
      }
    },
...
    "/projects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "autoApproveFeature": {
                        "type": "boolean"
                      },
                      "buildsCounter": {
                        "minimum": 0,
                        "type": "integer"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "diffThreshold": {
                        "type": [
                          "number",
                          "null"
                        ]
                      },
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "imageComparison": {
                        "enum": [
                          "odiff",
                          "pixelmatch",
                          "looks_same"
                        ],
                        "type": "string"
                      },
                      "imageComparisonConfig": {
                      },
                      "l2Enabled": {
                        "type": "boolean"
                      },
                      "mainBranchName": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "maxBranchLifetime": {
                        "minimum": 0,
                        "type": [
                          "integer",
                          "null"
                        ]
                      },
                      "maxBuildAllowed": {
                        "minimum": 0,
                        "type": [
                          "integer",
                          "null"
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "retentionDays": {
                        "minimum": 0,
                        "type": [
                          "integer",
                          "null"
                        ]
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "mainBranchName",
                      "buildsCounter",
                      "maxBuildAllowed",
                      "maxBranchLifetime",
                      "autoApproveFeature",
                      "imageComparison",
                      "retentionDays",
                      "diffThreshold",
                      "l2Enabled",
                      "createdAt",
                      "updatedAt"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Projects the caller has access to (admin: all; editor: member-of; guest: empty)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "List projects visible to the caller",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "mainBranchName": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "autoApproveFeature": {
                      "type": "boolean"
                    },
                    "buildsCounter": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "diffThreshold": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "imageComparison": {
                      "enum": [
                        "odiff",
                        "pixelmatch",
                        "looks_same"
                      ],
                      "type": "string"
                    },
                    "imageComparisonConfig": {
                    },
                    "l2Enabled": {
                      "type": "boolean"
                    },
                    "mainBranchName": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "maxBranchLifetime": {
                      "minimum": 0,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "maxBuildAllowed": {
                      "minimum": 0,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "retentionDays": {
                      "minimum": 0,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "mainBranchName",
                    "buildsCounter",
                    "maxBuildAllowed",
                    "maxBranchLifetime",
                    "autoApproveFeature",
                    "imageComparison",
                    "retentionDays",
                    "diffThreshold",
                    "l2Enabled",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Project created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not admin"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project name taken"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "Create a project (admin only)",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "autoApproveFeature": {
                      "type": "boolean"
                    },
                    "buildsCounter": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "diffThreshold": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "imageComparison": {
                      "enum": [
                        "odiff",
                        "pixelmatch",
                        "looks_same"
                      ],
                      "type": "string"
                    },
                    "imageComparisonConfig": {
                    },
                    "l2Enabled": {
                      "type": "boolean"
                    },
                    "mainBranchName": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "maxBranchLifetime": {
                      "minimum": 0,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "maxBuildAllowed": {
                      "minimum": 0,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "retentionDays": {
                      "minimum": 0,
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "mainBranchName",
                    "buildsCounter",
                    "maxBuildAllowed",
                    "maxBranchLifetime",
                    "autoApproveFeature",
                    "imageComparison",
                    "retentionDays",
                    "diffThreshold",
                    "l2Enabled",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Project record"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not a project member"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "Fetch a project by id (project members only)",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}